+++ affinda/models/_models_py3.py
@@ -535,6 +535,62 @@ class BaseExtractor(msrest.serialization.Model):
         self.created_dt = created_dt
 
 
+class BatchAddTagRequest(msrest.serialization.Model):
+    """BatchAddTagRequest.
+
+    :ivar identifiers: List of documents to tag.
+    :vartype identifiers: list[str]
+    :ivar tag: The tag's ID.
+    :vartype tag: int
+    """
+
+    _attribute_map = {
+        "identifiers": {"key": "identifiers", "type": "[str]"},
+        "tag": {"key": "tag", "type": "int"},
+    }
+
+    def __init__(
+        self, *, identifiers: Optional[List[str]] = None, tag: Optional[int] = None, **kwargs
+    ):
+        """
+        :keyword identifiers: List of documents to tag.
+        :paramtype identifiers: list[str]
+        :keyword tag: The tag's ID.
+        :paramtype tag: int
+        """
+        super(BatchAddTagRequest, self).__init__(**kwargs)
+        self.identifiers = identifiers
+        self.tag = tag
+
+
+class BatchRemoveTagRequest(msrest.serialization.Model):
+    """BatchRemoveTagRequest.
+
+    :ivar identifiers: List of documents to remove tag from.
+    :vartype identifiers: list[str]
+    :ivar tag: The tag's ID.
+    :vartype tag: int
+    """
+
+    _attribute_map = {
+        "identifiers": {"key": "identifiers", "type": "[str]"},
+        "tag": {"key": "tag", "type": "int"},
+    }
+
+    def __init__(
+        self, *, identifiers: Optional[List[str]] = None, tag: Optional[int] = None, **kwargs
+    ):
+        """
+        :keyword identifiers: List of documents to remove tag from.
+        :paramtype identifiers: list[str]
+        :keyword tag: The tag's ID.
+        :paramtype tag: int
+        """
+        super(BatchRemoveTagRequest, self).__init__(**kwargs)
+        self.identifiers = identifiers
+        self.tag = tag
+
+
 class Collection(msrest.serialization.Model):
     """Collection.
 
@@ -18136,6 +18192,9 @@ class RowAnnotation(Annotation):
 
     All required parameters must be populated in order to send to Azure.
 
+    :ivar additional_properties: Unmatched properties from the message are deserialized to this
+     collection.
+    :vartype additional_properties: dict[str, any]
     :ivar id: Required. Annotation's ID.
     :vartype id: int
     :ivar rectangle: Required. x/y coordinates for the rectangular bounding box containing the
@@ -18172,9 +18231,6 @@ class RowAnnotation(Annotation):
      "json", "table", "cell", "expectedremuneration", "jobtitle", "language", "skill",
      "yearsexperience", "group", "table_deprecated".
     :vartype content_type: str or ~affinda.models.AnnotationContentType
-    :ivar additional_properties: Unmatched properties from the message are deserialized to this
-     collection.
-    :vartype additional_properties: dict[str, any]
     :ivar parsed:
     :vartype parsed: ~affinda.models.RowAnnotationParsed
     """
@@ -18196,6 +18252,7 @@ class RowAnnotation(Annotation):
     }
 
     _attribute_map = {
+        "additional_properties": {"key": "", "type": "{object}"},
         "id": {"key": "id", "type": "int"},
         "rectangle": {"key": "rectangle", "type": "Rectangle"},
         "rectangles": {"key": "rectangles", "type": "[Rectangle]"},
@@ -18210,7 +18267,6 @@ class RowAnnotation(Annotation):
         "is_auto_verified": {"key": "isAutoVerified", "type": "bool"},
         "data_point": {"key": "dataPoint", "type": "str"},
         "content_type": {"key": "contentType", "type": "str"},
-        "additional_properties": {"key": "", "type": "{object}"},
         "parsed": {"key": "parsed", "type": "RowAnnotationParsed"},
     }
 
@@ -18230,12 +18286,15 @@ class RowAnnotation(Annotation):
         is_auto_verified: bool,
         data_point: str,
         content_type: Union[str, "_models.AnnotationContentType"],
-        document: Optional[str] = None,
         additional_properties: Optional[Dict[str, Any]] = None,
+        document: Optional[str] = None,
         parsed: Optional["_models.RowAnnotationParsed"] = None,
         **kwargs,
     ):
         """
+        :keyword additional_properties: Unmatched properties from the message are deserialized to this
+         collection.
+        :paramtype additional_properties: dict[str, any]
         :keyword id: Required. Annotation's ID.
         :paramtype id: int
         :keyword rectangle: Required. x/y coordinates for the rectangular bounding box containing the
@@ -18273,13 +18332,11 @@ class RowAnnotation(Annotation):
          "json", "table", "cell", "expectedremuneration", "jobtitle", "language", "skill",
          "yearsexperience", "group", "table_deprecated".
         :paramtype content_type: str or ~affinda.models.AnnotationContentType
-        :keyword additional_properties: Unmatched properties from the message are deserialized to this
-         collection.
-        :paramtype additional_properties: dict[str, any]
         :keyword parsed:
         :paramtype parsed: ~affinda.models.RowAnnotationParsed
         """
         super(RowAnnotation, self).__init__(
+            additional_properties=additional_properties,
             id=id,
             rectangle=rectangle,
             rectangles=rectangles,
@@ -18296,7 +18353,6 @@ class RowAnnotation(Annotation):
             content_type=content_type,
             **kwargs,
         )
-        self.additional_properties = additional_properties
         self.parsed = parsed
 
 
@@ -18322,7 +18378,7 @@ class RowAnnotationParsed(msrest.serialization.Model):
     :ivar item_tax_rate:
     :vartype item_tax_rate: ~affinda.models.TextAnnotation
     :ivar item_tax_total:
-    :vartype item_tax_total: ~affinda.models.TextAnnotation
+    :vartype item_tax_total: ~affinda.models.FloatAnnotation
     :ivar item_total:
     :vartype item_total: ~affinda.models.FloatAnnotation
     :ivar item_other:
@@ -18339,7 +18395,7 @@ class RowAnnotationParsed(msrest.serialization.Model):
         "item_discount": {"key": "itemDiscount", "type": "TextAnnotation"},
         "item_base_total": {"key": "itemBaseTotal", "type": "FloatAnnotation"},
         "item_tax_rate": {"key": "itemTaxRate", "type": "TextAnnotation"},
-        "item_tax_total": {"key": "itemTaxTotal", "type": "TextAnnotation"},
+        "item_tax_total": {"key": "itemTaxTotal", "type": "FloatAnnotation"},
         "item_total": {"key": "itemTotal", "type": "FloatAnnotation"},
         "item_other": {"key": "itemOther", "type": "TextAnnotation"},
     }
@@ -18356,7 +18412,7 @@ class RowAnnotationParsed(msrest.serialization.Model):
         item_discount: Optional["_models.TextAnnotation"] = None,
         item_base_total: Optional["_models.FloatAnnotation"] = None,
         item_tax_rate: Optional["_models.TextAnnotation"] = None,
-        item_tax_total: Optional["_models.TextAnnotation"] = None,
+        item_tax_total: Optional["_models.FloatAnnotation"] = None,
         item_total: Optional["_models.FloatAnnotation"] = None,
         item_other: Optional["_models.TextAnnotation"] = None,
         **kwargs,
@@ -18381,7 +18437,7 @@ class RowAnnotationParsed(msrest.serialization.Model):
         :keyword item_tax_rate:
         :paramtype item_tax_rate: ~affinda.models.TextAnnotation
         :keyword item_tax_total:
-        :paramtype item_tax_total: ~affinda.models.TextAnnotation
+        :paramtype item_tax_total: ~affinda.models.FloatAnnotation
         :keyword item_total:
         :paramtype item_total: ~affinda.models.FloatAnnotation
         :keyword item_other: